+++ UnitTests/OIDRegistrationResponseTests.m
@@ -20,8 +20,13 @@
 
 #import "OIDClientMetadataParameters.h"
 #import "OIDRegistrationRequestTests.h"
-#import "Source/OIDRegistrationRequest.h"
-#import "Source/OIDRegistrationResponse.h"
+
+#if SWIFT_PACKAGE
+@import AppAuthCore;
+#else
+#import "Source/AppAuthCore/OIDRegistrationRequest.h"
+#import "Source/AppAuthCore/OIDRegistrationResponse.h"
+#endif
 
 /*! @brief The test value for the @c clientID property.
  */